refactor(config): move lua/common/* to lua/config/* (LazyVim layout)
Pure relocation/rename, no behavior change:
- common/keymap.lua      -> config/keymaps.lua
- common/set_config.lua  -> config/options.lua
- common/autocmd_config.lua -> config/autocmds.lua
- common/{claude,server,treesitter_compat}.lua -> config/ (names kept)

Follow up all require/path references in init.lua, plugins/ui.lua,
plugins/claude.lua (comment), tests/treesitter_compat.sh and setup.sh.

plugins/configs/ kept as-is: flattening into plugins/ would make lazy.nvim's
`{ import = "plugins" }` load those config functions as plugin specs, which
would change behavior.

Verified: headless init.lua load (options/keymaps/Server* cmds/treesitter_compat
all registered) and `bash tests/run.sh` exit 0.

Closes #50
