+++ .config/nvim/lua/config/claude.lua
