+++ .config/nvim/lua/plugins/claude.lua
@@ -1,7 +1,7 @@
 return {
 	{
 		"mitubaEX/claude_workflow.nvim",
-		-- Loaded on first :Claude* command. Keymaps in lua/common/claude.lua and
+		-- Loaded on first :Claude* command. Keymaps in lua/config/claude.lua and
 		-- the worktree+claude flow in plugins/configs/worktree.lua both go through
 		-- `require("claude_workflow")`, which lazy.nvim's require-hook also picks
 		-- up — so no explicit `keys = {...}` block needed.
