+++ setup.sh
@@ -5,5 +5,5 @@ ln -sf $(pwd)/.config/nvim/init.lua $HOME/.config/nvim/init.lua
 # lua
 mkdir -p $HOME/.config/nvim/lua/lsp
 mkdir -p $HOME/.config/nvim/lua/plugins/configs
-mkdir -p $HOME/.config/nvim/lua/common
+mkdir -p $HOME/.config/nvim/lua/config
 find .config -type f | xargs -I% ln -sf $(pwd)/% $HOME/%
